+++ src/Model/SprintReport/SprintReportSprint.php
@@ -7,17 +7,11 @@ class SprintReportSprint
     public string $id;
     public string $name;
     public SprintStateEnum $state;
-    public ?int $startDateTimestamp;
-    public ?int $endDateTimestamp;
-    public ?int $completedDateTimestamp;
 
-    public function __construct(string $id, string $name, SprintStateEnum $state, ?int $startDateTimestamp, ?int $endDateTimestamp, ?int $completedDateTimestamp)
+    public function __construct(string $id, string $name, SprintStateEnum $state)
     {
         $this->id = $id;
         $this->name = $name;
         $this->state = $state;
-        $this->startDateTimestamp = $startDateTimestamp;
-        $this->endDateTimestamp = $endDateTimestamp;
-        $this->completedDateTimestamp = $completedDateTimestamp;
     }
 }