+++ src/Service/DataProviderService.php
@@ -29,7 +29,6 @@ class DataProviderService
         protected readonly string $sprintNameRegex,
         protected readonly int $httpClientRetryDelayMs = 1000,
         protected readonly int $httpClientMaxRetries = 3,
-        private readonly DateTimeHelper $dateTimeHelper,
     ) {
     }
 
@@ -83,7 +82,6 @@ class DataProviderService
                     $this->weekGoalLow,
                     $this->weekGoalHigh,
                     $this->sprintNameRegex,
-                    $this->dateTimeHelper
                 );
                 break;
             default: